💄 Implement custom renderers for crumbs
