+++ static/app/components/events/breadcrumbs/breadcrumbsTimeline.tsx
@@ -1,7 +1,7 @@
 import {Fragment} from 'react';
 import styled from '@emotion/styled';
 
-import {openNavigateToExternalLinkModal} from 'sentry/actionCreators/modal';
+import BreadcrumbsItemContent from 'sentry/components/events/breadcrumbs/breadcrumbsItemContent';
 import {
   BREADCRUMB_TIMESTAMP_PLACEHOLDER,
   BreadcrumbIcon,
@@ -9,17 +9,10 @@ import {
   getBreadcrumbColorConfig,
   getBreadcrumbTitle,
 } from 'sentry/components/events/breadcrumbs/utils';
-import {Sql} from 'sentry/components/events/interfaces/breadcrumbs/breadcrumb/data/sql';
-import {StructuredData} from 'sentry/components/structuredEventData';
 import Timeline, {type TimelineItemProps} from 'sentry/components/timeline';
 import {t} from 'sentry/locale';
-import {
-  BreadcrumbMessageFormat,
-  BreadcrumbType,
-  type RawCrumb,
-} from 'sentry/types/breadcrumbs';
+import type {RawCrumb} from 'sentry/types/breadcrumbs';
 import {defined} from 'sentry/utils';
-import {isUrl} from 'sentry/utils/string/isUrl';
 
 interface BreadcrumbsTimelineProps {
   breadcrumbs: RawCrumb[];
@@ -73,114 +66,9 @@ export default function BreadcrumbsTimeline({
   return <Timeline.Container>{items}</Timeline.Container>;
 }
 
-interface BreadcrumbsItemContentProps {
-  breadcrumb: RawCrumb;
-  fullyExpanded?: boolean;
-  meta?: Record<string, any>;
-}
-
-function BreadcrumbsItemContent({
-  breadcrumb: bc,
-  meta,
-  fullyExpanded,
-}: BreadcrumbsItemContentProps) {
-  const maxDefaultDepth = fullyExpanded ? 10000 : 1;
-  const structureProps = {
-    depth: 0,
-    maxDefaultDepth,
-    withAnnotatedText: true,
-    withOnlyFormattedText: true,
-  };
-
-  const defaultMessage = defined(bc.message) ? (
-    <Timeline.Text>
-      <StructuredData value={bc.message} meta={meta?.message} {...structureProps} />
-    </Timeline.Text>
-  ) : null;
-  const defaultData = defined(bc.data) ? (
-    <Timeline.Data>
-      <StructuredData value={bc.data} meta={meta?.data} {...structureProps} />
-    </Timeline.Data>
-  ) : null;
-
-  if (bc?.type === BreadcrumbType.HTTP) {
-    const {method, url, status_code: statusCode, ...otherData} = bc.data;
-
-    return (
-      <Fragment>
-        {defaultMessage}
-        <Timeline.Text>
-          {method && `${method}: `}
-          {url && isUrl(url) ? (
-            <Link onClick={() => openNavigateToExternalLinkModal({linkText: url})}>
-              {url}
-            </Link>
-          ) : (
-            url
-          )}
-          {` [${statusCode}]`}
-        </Timeline.Text>
-        {defined(bc.data) ? (
-          <Timeline.Data>
-            <StructuredData value={otherData} meta={meta?.data} {...structureProps} />
-          </Timeline.Data>
-        ) : null}
-      </Fragment>
-    );
-  }
-
-  if (
-    !defined(meta) &&
-    bc?.message &&
-    bc?.messageFormat === BreadcrumbMessageFormat.SQL
-  ) {
-    return (
-      <Timeline.Data>
-        <LightenTextColor>
-          <Sql breadcrumb={bc} searchTerm="" />
-        </LightenTextColor>
-      </Timeline.Data>
-    );
-  }
-  if (
-    (!defined(meta) && bc?.type === BreadcrumbType.WARNING) ||
-    bc?.type === BreadcrumbType.ERROR
-  ) {
-    return (
-      <Fragment>
-        <Timeline.Text>
-          {bc?.data?.type && `${bc?.data?.type}: `}
-          {bc?.data?.value}
-        </Timeline.Text>
-        {defaultMessage}
-      </Fragment>
-    );
-  }
-
-  return (
-    <Fragment>
-      {defaultMessage}
-      {defaultData}
-    </Fragment>
-  );
-}
-
 const Subtitle = styled('p')`
   margin: 0;
   font-weight: normal;
   font-size: ${p => p.theme.fontSizeSmall};
   display: inline;
-`;
-
-const Link = styled('a')`
-  color: ${p => p.theme.subText};
-  text-decoration: underline;
-  text-decoration-style: dotted;
-  word-break: break-all;
-`;
-
-const LightenTextColor = styled('div')`
-  .token {
-    color: ${p => p.theme.subText};
-  }
 `;
